requirements.txt
sondera-harness[langgraph]>=0.6.3
langchain>=1.0.0
langchain-anthropic>=0.3.0
biopython>=1.84
python-dotenv>=1.0.0

# Web demo (web.py)
fastapi>=0.115.0
uvicorn>=0.30.0
